% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biobtreeR.R
\name{bbAttr}
\alias{bbAttr}
\title{Retrieve attributes of dataset}
\usage{
bbAttr(id, source)
}
\arguments{
\item{source}{Dataset identifier}

\item{identifer}{Identifer for the entry.}
}
\description{
Provides list of available attributes for a dataset to use in search and mapping queries. This function works with a sample identifer.
}
\examples{
bbAttr("p15498","uniprot")
}
\author{
Tamer Gür
}
