% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildData.R
\name{bbBuildCustomDB}
\alias{bbBuildCustomDB}
\title{Build custom DB}
\usage{
bbBuildCustomDB(taxonomyIDs = NULL, rawArgs = NULL)
}
\arguments{
\item{taxonomyIDs}{Comma seperated list of taxonomy identifiers for building the genomes}

\item{rawArgs}{For using all available biobtree command line arguments directly}
}
\value{
returns empty
}
\description{
biobtree covers all the genomes in ensembl and ensembl genomes. If the the studied organism genome is not included
in the default pre built in databases then this function is used and build the biobtree database locally for given genomes.
}
\examples{

\dontrun{

bbUseOutDir("your directory path")
bbBuildCustomDB(taxonomyIDs="1408103,206403")

}

}
\author{
Tamer Gur
}
