% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildData.R
\name{bbBuiltInDB}
\alias{bbBuiltInDB}
\title{Get pre build biobtree database}
\usage{
bbBuiltInDB(type = "1")
}
\arguments{
\item{type}{built in database type accepted values are 1,2 and 3. Currently there are 3 different builtin database;
Type 1
Included datasets hgnc,hmdb,taxonomy,go,efo,eco,chebi,interpro
Included uniprot proteins and ensembl genomes belongs to following organisms
homo_sapiens 9606 --> ensembl
danio_rerio 7955 zebrafish --> ensembl
gallus_gallus 9031 chicken --> ensembl
mus_musculus 10090 --> ensembl
Rattus norvegicus 10116 ---> ensembl
saccharomyces_cerevisiae 4932--> ensembl
arabidopsis_thaliana 3702--> ensembl
drosophila_melanogaster 7227 --> ensembl
caenorhabditis_elegans 6239 --> ensembl
Escherichia coli 562 --> ensembl_bacteria
Escherichia coli str. K-12 substr. MG1655 511145 --> ensembl_bacteria
Escherichia coli K-12 83333 --> ensembl_bacteria

Type 2
Instead of genomes in in the type 1 it contains human and all the mouse strains genomes in the ensembl

Type 3
Contains no genome but it contains all the uniprot data.}
}
\value{
returns empty
}
\description{
Pre build biobtree database for commonly studied datasets and model organism genomes. Once this function called it retrieves
the pre build database saves to users output directory.
}
\examples{

bbUseOutDir(tempdir()) # temp dir for demo purpose
bbBuiltInDB("demo") # small demo database for real database use 1, 2 or 3


}
\author{
Tamer Gur
}
