% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biobtreeR.R
\name{bbEntry}
\alias{bbEntry}
\title{Retrieve entry}
\usage{
bbEntry(identifer, source)
}
\arguments{
\item{identifer}{Identifer for the entry. Note that keywords are not accepted. For instance insted of "vav_human" keyword "p15498" identifier must be passed}

\item{source}{Dataset identifier}
}
\description{
Returns entry for an identifier and dataset. Entry contains all the data raw data for and entry such as mappings, attiributes and paging info if exists.
}
\examples{
bbEntry("p15498","uniprot")
}
\author{
Tamer Gür
}
