% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biobtreeR.R
\name{bbEntryFilter}
\alias{bbEntryFilter}
\title{Retrieve entry with filtered dataset}
\usage{
bbEntryFilter(id, source, filters, page = NULL)
}
\arguments{
\item{source}{Dataset identifier}

\item{filters}{Comma seperated dataset identifer to retrieve}

\item{page}{Page index if results is more than default biobtree paging size.}

\item{identifer}{Identifer for the entry.}
}
\description{
Similar with entry retrieval but filtered mapping entries with given datasets.
}
\examples{
bbEntryFilter("p15498","uniprot","go,kegg")
}
\author{
Tamer Gür
}
