% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biobtreeR.R
\name{bbEntryPage}
\alias{bbEntryPage}
\title{Retrieve entry result page}
\usage{
bbEntryPage(identifer, source, page, totalPage)
}
\arguments{
\item{identifer}{Identifer for the entry.}

\item{source}{Dataset identifier}

\item{page}{Page index it starts from 0}

\item{totalPage}{Total number of page for the entry. This value needs to calculate by user via using total number of entries which is available at the root result for the entry
and divide it confiGüred biobtree paging size which has default value of 200}
}
\description{
If an entry contains large set of mapping entries it is paginated by biobtree with confiGüred paging size. This function retrieve these paging for an entry
}
\examples{
bbEntryPage("p15498","uniprot",0,6)
}
\author{
Tamer Gür
}
