% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biobtreeR.R
\name{bbInit}
\alias{bbInit}
\title{Init function for biobtreeR}
\usage{
bbInit(biobtreeURL = NULL)
}
\arguments{
\item{biobtreeURL}{Optional parameter. By default "http://localhost:8888" url is used to connect running actual biobtree.
This parameter is allowed to change this url if biobtree is running from another machine or port.}
}
\value{
returns empty
}
\description{
Initialize all dataset meta information such as dataset identifiers.
In biobtree each dataset has unique charachter and numeric identifier.
For instance Uniprot charachter identifier is "uniprot" and numeric identifier is 1.
When querying the dataset charachter identifier is used for convinience but in actual database it is saved numerically for efficiency.
When this function is executed global bbConfig variable is set. This variable holds these meta information.
}
\examples{
bbInit()
}
\author{
Tamer Gur
}
