% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biobtreeR.R
\name{bbListDatasets}
\alias{bbListDatasets}
\title{List available datasets}
\usage{
bbListDatasets(outDir = NULL, biobtreeURL = NULL)
}
\arguments{
\item{outDir}{If outDir specified during bbBuiltData specified same one here. If not specified tempdir is used.}

\item{biobtreeURL}{Use this this parameter if you running biobtree tool seperately such as in a remote server.
Then set this parameter with the endpoint of this running biobtree url with its ip address and port http://${IP}:${PORT} .
Note that default port of biobtree is 8888.}
}
\value{
returns list of datasets
}
\description{
This function list the available source and target datasets with their numeric identifiers. If biobtree web server
is not runnint this function triggers to run it. And also if no data built with bbBuildData before it create a example data using
hgnc just to show the available datasets since biobtree web server require at least one built data to properly start.
}
\examples{

bbListDatasets()

}
