% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{bbStart}
\alias{bbStart}
\title{Start biobtreeR}
\usage{
bbStart(outDir = NULL)
}
\arguments{
\item{outDir}{If outDir specified during bbBuiltData specified same one here. If not specified tempdir is used.}
}
\value{
character
}
\description{
Once target datasets is built with \code{bbBuildData} this function used to start biobtree server
in the background for performing search/mapping queries.
}
\examples{
bbStart()
bbStop()
}
