% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{bbUseOutDir}
\alias{bbUseOutDir}
\title{Output directory for biobtreeR}
\usage{
bbUseOutDir(outDir)
}
\arguments{
\item{outDir}{path for the output directory.}
}
\value{
returns empty
}
\description{
Allows to set the directory for the package for its files. It is required to set a valid directory.
}
\examples{

bbUseOutDir(tempdir())

}
