% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{bbUseRemoteServer}
\alias{bbUseRemoteServer}
\title{Use remote server}
\usage{
bbUseRemoteServer(remoteEndpoint)
}
\arguments{
\item{remoteEndpoint}{remote endpoint url with ip address and port in the form of http://${IP}:${PORT} .
Note that default port of biobtree is 8888.}
}
\value{
empty
}
\description{
Use this function if you running biobtree tool seperately in a remote server
}
