% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biobtreeR.R
\name{bbAttr}
\alias{bbAttr}
\title{Retrieve attributes of dataset}
\usage{
bbAttr(identifer, source)
}
\arguments{
\item{identifer}{Identifer for the entry.}

\item{source}{Dataset identifier}
}
\value{
returns attributes names
}
\description{
Provides list of available attributes for a dataset to use in search and mapping queries. This function works with a sample identifer.
}
\author{
Tamer Gur
}
