% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{bbListDatasets}
\alias{bbListDatasets}
\title{List available datasets}
\usage{
bbListDatasets()
}
\value{
returns datasets
}
\description{
Lists the available source and target datasets with their numeric identifiers.
}
\examples{

bbListDatasets()

}
