% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biobtreeR.R
\name{bbListGenomes}
\alias{bbListGenomes}
\title{List Genomes}
\usage{
bbListGenomes(ensemblType)
}
\arguments{
\item{ensemblType}{These param can be one of the following "ensembl", "ensembl_bacteria", "ensembl_fungi", "ensembl_metazoa", "ensembl_plants", "ensembl_protists"}
}
\value{
returns list of genome names
}
\description{
This function list the ensembl genome names. These names can be used in bbBuildData function
}
\examples{

bbStart()
bbListGenomes("ensembl")
#For ensembl genomes
bbListGenomes("ensembl_bacteria")
bbListGenomes("ensembl_fungi")
bbListGenomes("ensembl_metazoa")
bbListGenomes("ensembl_plants")
bbListGenomes("ensembl_protists")

}
