% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biobtreeR.R
\name{bbStart}
\alias{bbStart}
\title{Start biobtreeR}
\usage{
bbStart(biobtreeURL = NULL)
}
\arguments{
\item{biobtreeURL}{Optional parameter. Use this this parameter if you running biobtree tool seperately such as in a remote server.
Then set this parameter with the endpoint of this running biobtree url with its ip address and port http://${IP}:${PORT} .
Note that default port of biobtree is 8888.}
}
\value{
returns empty
}
\description{
Once target datasets is built with \code{bbBuildData} this function used to start biobtree server in the background for performing search/mapping queries.
This function also sets the bbConfig variable
}
\examples{
bbStart()
bbStop()
}
\author{
Tamer Gur
}
