% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biobtreeR.R
\name{bbURL}
\alias{bbURL}
\title{Retrieve unique url of entry}
\usage{
bbURL(identifer, source)
}
\arguments{
\item{identifer}{Identifer for the entry.}

\item{source}{Dataset identifier}
}
\value{
returns unique url

 @author Tamer Gur
}
\description{
Provides unique url for the identifier and dataset
}
\examples{
bbStart()
bbURL("HGNC:12009","hgnc")
bbURL("p15498","uniprot")
bbStop()

}
