% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biobtreeR.R
\name{bbURL}
\alias{bbURL}
\title{Retrieve unique url of entry}
\usage{
bbURL(id, source)
}
\arguments{
\item{source}{Dataset identifier}

\item{identifer}{Identifer for the entry.}
}
\description{
Provides unique url for the identifier and dataset
}
\examples{
bbURL("p15498","uniprot")
}
\author{
Tamer Gür
}
